combine test sites
Combines the browser-based test sites for the contained packages
into a single directory, with testing logic for both of the
libraries contained in it. The node-related tests will remain
in chat-core explicitly, as chat-core-aws-connect does not
support Node.
