+++ packages/chat-core-aws-connect/src/models/connect.d.ts
@@ -8,4 +8,5 @@ declare global {
         export type ActiveCustomerChatSession = Extract<ActiveChatSession, { disconnectParticipant(): unknown }>;
         export type Logger = typeof AwsChatSession.Logger;
     }
+    const AWS: unknown;
 }
