--- packages/chat-core-aws-connect/test-browser-esm/index.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <title>Chat Core Test Site</title>
-    <script type="module" src="./dist/bundle.js"></script>
-  </head>
-  <body>
-    <input
-      id="messageInput"
-      type="text"
-      value="I would like to speak to an agent"
-    /><button onclick="getNextMessage()">Chat</button>
-    <p id="chatresult"></p>
-    <ul id="messages"></ul>
-    <button onclick="resetSession()">Reset Session</button>
-  </body>
-</html>
