--- packages/chat-core/test-browser-esm/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <title>Chat Core Test Site</title>
-    <script type="module" src="./dist/bundle.js"></script>
-  </head>
-  <body>
-    <button onclick="getNextMessage()">Chat</button>
-    <p id="chatresult"></p>
-    <button onclick="streamNextMessage()">Chat (Stream)</button>
-    <p id="chatresult-stream"></p>
-  </body>
-</html>
