--- packages/chat-core/test-browser-esm/sample.env
@@ -1,2 +1,0 @@
-TEST_BOT_ID=
-TEST_BOT_API_KEY=
